+++ packages/api/src/router/user.ts
@@ -175,7 +175,7 @@ export const userRouter = createTRPCRouter({
   // Is protected because also used in board access / integration access forms
   selectable: protectedProcedure
     .input(z.object({ excludeExternalProviders: z.boolean().default(false) }).optional())
-    .output(z.array(selectUserSchema.pick({ id: true, name: true, image: true })))
+    .output(z.array(selectUserSchema.pick({ id: true, name: true, image: true, email: true })))
     .meta({ openapi: { method: "GET", path: "/api/users/selectable", tags: ["users"], protect: true } })
     .query(({ ctx, input }) => {
       return ctx.db.query.users.findMany({
@@ -183,6 +183,7 @@ export const userRouter = createTRPCRouter({
           id: true,
           name: true,
           image: true,
+          email: true,
         },
         where: input?.excludeExternalProviders ? eq(users.provider, "credentials") : undefined,
       });
@@ -195,7 +196,7 @@ export const userRouter = createTRPCRouter({
         limit: z.number().min(1).max(100).default(10),
       }),
     )
-    .output(z.array(selectUserSchema.pick({ id: true, name: true, image: true })))
+    .output(z.array(selectUserSchema.pick({ id: true, name: true, image: true, email: true })))
     .meta({ openapi: { method: "POST", path: "/api/users/search", tags: ["users"], protect: true } })
     .query(async ({ input, ctx }) => {
       const dbUsers = await ctx.db.query.users.findMany({
@@ -203,6 +204,7 @@ export const userRouter = createTRPCRouter({
           id: true,
           name: true,
           image: true,
+          email: true,
         },
         where: like(users.name, `%${input.query}%`),
         limit: input.limit,
@@ -211,6 +213,7 @@ export const userRouter = createTRPCRouter({
         id: user.id,
         name: user.name ?? "",
         image: user.image,
+        email: user.email,
       }));
     }),
   getById: protectedProcedure