+++ packages/integrations/src/unraid/unraid-integration.ts
@@ -67,6 +67,8 @@ export class UnraidIntegration extends Integration implements ISystemHealthMonit
         deviceName: disk.name,
         temperature: disk.temp ?? null,
         overallStatus: disk.status,
+        // See ArrayDiskStatus from https://studio.apollographql.com/public/Unraid-API/variant/current/explorer
+        healthy: disk.status === "DISK_OK",
       })),
     };
   }